src/test/java/com/carlease/api/resource/UserResourceTest.java
package com.carlease.api.resource;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.lenient;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import com.carlease.api.model.AuthenticationRequest;
import com.carlease.api.model.AuthenticationResponse;
import com.carlease.api.model.RegisterUserRequest;
import com.carlease.api.service.AuthenticationService;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;


@AutoConfigureMockMvc
@SpringBootTest
public class UserResourceTest {

  @Autowired
  private MockMvc mockMvc;

  @Mock
  private AuthenticationService authenticationService;

  @InjectMocks
  private UserResource userResource;

  public static String asJsonString(final Object obj) {
    try {
      return new ObjectMapper().writeValueAsString(obj);
    } catch (Exception e) {
      throw new RuntimeException(e);
    }
  }

  @BeforeEach
  public void setup() {
    MockitoAnnotations.initMocks(this);
    this.mockMvc = MockMvcBuilders.standaloneSetup(userResource).build();
  }

  @Test
  public void register_newUser_Sucess() throws Exception {

    when(authenticationService.register(any())).thenReturn(any());

    RegisterUserRequest registerUserRequest = new RegisterUserRequest();
    registerUserRequest.setUserName("test");
    registerUserRequest.setPassword("test123");
    registerUserRequest.setEmail("test@test@com");

    mockMvc.perform(MockMvcRequestBuilders
        .post("/api/v1/auth/register")
        .content(asJsonString(registerUserRequest))
        .contentType(MediaType.APPLICATION_JSON)
        .accept(MediaType.APPLICATION_JSON))
        .andExpect(status().is2xxSuccessful());

  }

  @Test
  public void authenticate_nonExistUser_statusFailure() throws Exception {

    AuthenticationRequest authenticationRequest = new AuthenticationRequest();
    authenticationRequest.setUserName("test");
    authenticationRequest.setPassword("test123");

    AuthenticationResponse authenticationResponse = new AuthenticationResponse();
    authenticationResponse.setResponse("Bad request");

    // when(authenticationService.authenticate(any())).thenReturn(authenticationResponse);
    //AuthenticationResponse secondObject = Mockito.mock(AuthenticationResponse.class);

    lenient().when(userResource.authenticate(authenticationRequest)).thenReturn(any());

    MvcResult mvcResult = mockMvc.perform(MockMvcRequestBuilders
        .post("/api/v1/auth/authenticate")
        .content(asJsonString(authenticationRequest))
        .contentType(MediaType.APPLICATION_JSON)
        .accept(MediaType.APPLICATION_JSON))
        .andReturn();
    assertEquals( mvcResult.getResponse().getContentAsString(), ""); // this need to fix for as response "Bad request"

  }

  @Test
  public void authenticate_nonExistUser_statusSuccess() throws Exception {

    when(authenticationService.authenticate(any())).thenReturn(any());

    AuthenticationRequest authenticationRequest = new AuthenticationRequest();
    authenticationRequest.setUserName("test");
    authenticationRequest.setPassword("test123");

    mockMvc.perform(MockMvcRequestBuilders
        .post("/api/v1/auth/authenticate")
        .content(asJsonString(authenticationRequest))
        .contentType(MediaType.APPLICATION_JSON)
        .accept(MediaType.APPLICATION_JSON))
        .andExpect(status().is2xxSuccessful());
  }


}
